API/requirements.txt
flask==3.1.0
tensorflow==2.19.0
pillow
numpy
werkzeug
gunicorn
boto3